+++ iOS/buildInformation.json
@@ -1,1 +1,1 @@
-{"gitSha": "ef485992f649db1b4b583e64f807ed4eb31c9512", "numberOfCommits": "332"}
+{"gitSha": "5a5567ba93da96e7e2007d928fc08cfcabadcea2", "numberOfCommits": "329"}
